requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
dj-database-url==1.0.0
Django==4.1.3
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==7.0.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
idna==3.4
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
urllib3==1.26.13
whitenoise==6.2.0
